+++ tests/Acme81/Attribute/SampleComplex.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Acme81\Attribute;
+
+use Attribute;
+
+#[Attribute(Attribute::TARGET_CLASS)]
+class SampleComplex
+{
+    public function __construct(
+        public SampleComplexValue $value,
+    ) {
+    }
+}